+++ src/java/modelo/DetallePedido.java
@@ -13,59 +13,59 @@ import java.util.ArrayList;
  */
 public class DetallePedido {
 
-    private double id_pedido;
-    private double id_producto;
-    private double cantidad;
+    private long id_pedido;
+    private long id_producto;
+    private long cantidad;
     
-    private ArrayList<Double> id_pedido_array = new ArrayList<>();
-    private ArrayList<Double> id_producto_array = new ArrayList<>();
-    private ArrayList<Double> cantidad_array = new ArrayList<>();
+    private ArrayList<Long> id_pedido_array = new ArrayList<>();
+    private ArrayList<Long> id_producto_array = new ArrayList<>();
+    private ArrayList<Long> cantidad_array = new ArrayList<>();
 
-    public double getId_pedido() {
+    public long getId_pedido() {
         return id_pedido;
     }
 
-    public void setId_pedido(double id_pedido) {
+    public void setId_pedido(long id_pedido) {
         this.id_pedido = id_pedido;
     }
 
-    public double getId_producto() {
+    public long getId_producto() {
         return id_producto;
     }
 
-    public void setId_producto(double id_producto) {
+    public void setId_producto(long id_producto) {
         this.id_producto = id_producto;
     }
-    
-    public double getCantidad() {
+
+    public long getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(double cantidad) {
+    public void setCantidad(long cantidad) {
         this.cantidad = cantidad;
     }
 
-    public ArrayList<Double> getId_pedido_array() {
+    public ArrayList<Long> getId_pedido_array() {
         return id_pedido_array;
     }
 
-    public void setId_pedido_array(ArrayList<Double> id_pedido_array) {
+    public void setId_pedido_array(ArrayList<Long> id_pedido_array) {
         this.id_pedido_array = id_pedido_array;
     }
 
-    public ArrayList<Double> getId_producto_array() {
+    public ArrayList<Long> getId_producto_array() {
         return id_producto_array;
     }
 
-    public void setId_producto_array(ArrayList<Double> id_producto_array) {
+    public void setId_producto_array(ArrayList<Long> id_producto_array) {
         this.id_producto_array = id_producto_array;
     }
 
-    public ArrayList<Double> getCantidad_array() {
+    public ArrayList<Long> getCantidad_array() {
         return cantidad_array;
     }
 
-    public void setCantidad_array(ArrayList<Double> cantidad_array) {
+    public void setCantidad_array(ArrayList<Long> cantidad_array) {
         this.cantidad_array = cantidad_array;
     }
